+++ tests/mersenne61.rs
@@ -81,3 +81,69 @@ fn mult_conmutativity() {
         assert_eq!(mult1, mult2);
     }
 }
+
+use proptest::prelude::*;
+
+mod common;
+
+fn element() -> impl Strategy<Value = Mersenne61> {
+    common::field_element()
+}
+
+proptest! {
+    #[test]
+    fn mul_distributes_over_add(a in element(), b in element(), c in element()) {
+        prop_assert_eq!(a * &(b + &c), (a * &b) + &(a * &c));
+    }
+
+    #[test]
+    fn mul_inverse_equals_one(a in element()) {
+        prop_assume!(a != Mersenne61::ZERO);
+        prop_assert_eq!(a * &(a.inverse().unwrap()), Mersenne61::ONE)
+    }
+
+    #[test]
+    fn add_commutes(a in element(), b in element()) {
+        prop_assert_eq!(a + &b, b + &a);
+    }
+
+    #[test]
+    fn add_associates(a in element(), b in element(), c in element()) {
+        prop_assert_eq!((a + &b) + &c, a + &(b + &c));
+    }
+
+    #[test]
+    fn add_identity(a in element()) {
+        prop_assert_eq!(a + &Mersenne61::ZERO, a);
+    }
+
+    #[test]
+    fn add_inverse_equals_zero(a in element()) {
+        prop_assert_eq!(a + &a.negate(), Mersenne61::ZERO);
+    }
+
+    #[test]
+    fn mul_commutes(a in element(), b in element()) {
+        prop_assert_eq!(a * &b, b * &a);
+    }
+
+    #[test]
+    fn mul_associates(a in element(), b in element(), c in element()) {
+        prop_assert_eq!((a * &b) * &c, a * &(b * &c));
+    }
+
+    #[test]
+    fn mul_identity(a in element()) {
+        prop_assert_eq!(a * &Mersenne61::ONE, a);
+    }
+
+    #[test]
+    fn sub_self_equals_zero(a in element()) {
+        prop_assert_eq!(a - &a, Mersenne61::ZERO);
+    }
+
+    #[test]
+    fn sub_equals_add_negate(a in element(), b in element()) {
+        prop_assert_eq!(a - &b, a + &b.negate());
+    }
+}