+++ tests/scalar_secp256k1.rs
@@ -27,3 +27,76 @@ fn addition_of_one_gives_two() {
     let two = Secp256k1ScalarField::from(2);
     assert_eq!(sum, two);
 }
+
+use proptest::prelude::*;
+
+use crate::common::roundtrip;
+
+mod common;
+
+fn element() -> impl Strategy<Value = Secp256k1ScalarField> {
+    common::field_element()
+}
+
+proptest! {
+    #[test]
+    fn mul_distributes_over_add(a in element(), b in element(), c in element()) {
+        prop_assert_eq!(a * &(b + &c), (a * &b) + &(a * &c));
+    }
+
+    #[test]
+    fn mul_inverse_equals_one(a in element()) {
+        prop_assume!(a != Secp256k1ScalarField::ZERO);
+        prop_assert_eq!(a * &(a.inverse().unwrap()), Secp256k1ScalarField::ONE)
+    }
+
+    #[test]
+    fn add_commutes(a in element(), b in element()) {
+        prop_assert_eq!(a + &b, b + &a);
+    }
+
+    #[test]
+    fn add_associates(a in element(), b in element(), c in element()) {
+        prop_assert_eq!((a + &b) + &c, a + &(b + &c));
+    }
+
+    #[test]
+    fn add_identity(a in element()) {
+        prop_assert_eq!(a + &Secp256k1ScalarField::ZERO, a);
+    }
+
+    #[test]
+    fn add_inverse_equals_zero(a in element()) {
+        prop_assert_eq!(a + &a.negate(), Secp256k1ScalarField::ZERO);
+    }
+
+    #[test]
+    fn mul_commutes(a in element(), b in element()) {
+        prop_assert_eq!(a * &b, b * &a);
+    }
+
+    #[test]
+    fn mul_associates(a in element(), b in element(), c in element()) {
+        prop_assert_eq!((a * &b) * &c, a * &(b * &c));
+    }
+
+    #[test]
+    fn mul_identity(a in element()) {
+        prop_assert_eq!(a * &Secp256k1ScalarField::ONE, a);
+    }
+
+    #[test]
+    fn sub_self_equals_zero(a in element()) {
+        prop_assert_eq!(a - &a, Secp256k1ScalarField::ZERO);
+    }
+
+    #[test]
+    fn sub_equals_add_negate(a in element(), b in element()) {
+        prop_assert_eq!(a - &b, a + &b.negate());
+    }
+
+    #[test]
+    fn postcard_roundtrip(a in element()) {
+        roundtrip(a)?;
+    }
+}